api/pyproject.toml
[project]
name = "hrclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "flask-admin>=1.6.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "sqlalchemy-utils>=0.41.2",
    # https://github.com/astral-sh/rye/issues/745
    "gunicorn>=22.0.0",
    "gevent>=24.2.1",
    "flask-cors>=5.0.0",
    "flask-pydantic>=0.12.0",
    "pydantic-settings>=2.6.1",
    "flask-compress>=1.17",
    "pytz>=2024.2",
    "email-validator>=2.2.0",
]

[dependency-groups]
dev = ["pytest-mock>=3.14.0", "pytest>=8.3.4", "yapf>=0.43.0"]

[tool.pytest.ini_options]
pythonpath = "."

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
